Fixed a few logic bugs

+++ LouieDrewHW10/doctor.h
@@ -9,7 +9,7 @@
 
 using namespace std;
 
-const int MAX_OZ_SCHRAUT = 25;
+const short MAX_OZ_SCHRAUT = 25;
 
 // Constructor sets m_name from parameter, sets m_oz_schraut to 0, and m_money
 //   to 0.
